+++ src/main/java/Main.java
@@ -3,7 +3,7 @@ public class Main {
     public static void main(String[] args) {
 
 // класс калькулятора
-        Calculate calculate=new Calculate();
+        Calculator calculate=new Calculator();
         calculate.qPerson();
         calculate.addProducts();
